renamed response formatter. Added support for selecting raw perm strings dynamically.
