+++ api/mappings/permissions/raw_permissions.py
@@ -1,0 +1,11 @@
+
+raw_perm_mappings = {
+                        'team':
+                        {
+                            'get': 'pyteam.retrieve_team',
+                            'create': 'pyteams.add_team',
+                            'update': 'pyteam.change_team',
+                            'delete': 'pyteam.delete_team',
+                        }
+
+                    }
